+++ src/search-strategies/benchmarking/indexOf-knuth-morris-pratt/search-strategy.ts
@@ -52,8 +52,7 @@ export class IndexOfKnuthMorrisPrattSearchStrategy
             yield {
               isMatch: true,
               content: match,
-              startIndex: matchStart,
-              endIndex: matchStart + this.needle.length
+              streamIndices: [matchStart, matchStart + this.needle.length]
             };
           }
         } else {
@@ -93,8 +92,7 @@ export class IndexOfKnuthMorrisPrattSearchStrategy
         yield {
           isMatch: true,
           content: this.needle,
-          startIndex,
-          endIndex
+          streamIndices: [startIndex, endIndex]
         };
       }
     } finally {